+++ src/main/java/net/geforcemods/securitycraft/SCClientEventHandler.java
@@ -1,10 +1,16 @@
 package net.geforcemods.securitycraft;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
 import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.function.Supplier;
 
+import org.lwjgl.glfw.GLFW;
+
+import com.mojang.blaze3d.pipeline.RenderTarget;
 import com.mojang.blaze3d.platform.Window;
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
@@ -14,44 +20,66 @@ import net.geforcemods.securitycraft.blockentities.BlockChangeDetectorBlockEntit
 import net.geforcemods.securitycraft.blockentities.BlockChangeDetectorBlockEntity.ChangeEntry;
 import net.geforcemods.securitycraft.blockentities.SecurityCameraBlockEntity;
 import net.geforcemods.securitycraft.blocks.SecurityCameraBlock;
+import net.geforcemods.securitycraft.entity.camera.CameraController;
+import net.geforcemods.securitycraft.entity.camera.CameraController.CameraFeed;
+import net.geforcemods.securitycraft.entity.camera.CameraViewAreaExtension;
 import net.geforcemods.securitycraft.misc.BlockEntityTracker;
 import net.geforcemods.securitycraft.misc.CameraRedstoneModuleState;
 import net.geforcemods.securitycraft.misc.KeyBindings;
 import net.geforcemods.securitycraft.misc.ModuleType;
 import net.geforcemods.securitycraft.util.Utils;
+import net.minecraft.client.Camera;
+import net.minecraft.client.CameraType;
 import net.minecraft.client.DeltaTracker;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.Options;
 import net.minecraft.client.gui.Font;
 import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.client.renderer.LevelRenderer;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.ShapeRenderer;
+import net.minecraft.client.renderer.chunk.SectionRenderDispatcher;
+import net.minecraft.client.renderer.chunk.SectionRenderDispatcher.RenderSection;
+import net.minecraft.client.renderer.culling.Frustum;
 import net.minecraft.core.BlockPos;
+import net.minecraft.core.GlobalPos;
 import net.minecraft.core.component.DataComponentType;
 import net.minecraft.core.component.DataComponents;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.ARGB;
+import net.minecraft.util.Mth;
+import net.minecraft.util.profiling.Profiler;
+import net.minecraft.util.profiling.ProfilerFiller;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.effect.MobEffects;
+import net.minecraft.world.entity.Entity;
+import net.minecraft.world.entity.EntityType;
+import net.minecraft.world.entity.Marker;
+import net.minecraft.world.entity.Pose;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item.TooltipContext;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.component.TooltipDisplay;
 import net.minecraft.world.item.component.TooltipProvider;
+import net.minecraft.world.level.ChunkPos;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.Vec3;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.neoforge.client.event.InputEvent;
+import net.neoforged.neoforge.client.event.RenderFrameEvent;
 import net.neoforged.neoforge.client.event.RenderHandEvent;
 import net.neoforged.neoforge.client.event.RenderLevelStageEvent;
 import net.neoforged.neoforge.client.event.RenderLevelStageEvent.Stage;
 import net.neoforged.neoforge.event.entity.player.ItemTooltipEvent;
+import net.neoforged.neoforge.event.level.ChunkEvent;
 import net.neoforged.neoforge.registries.DeferredHolder;
 
 @EventBusSubscriber(modid = SecurityCraft.MODID, value = Dist.CLIENT)
@@ -132,6 +160,179 @@ public class SCClientEventHandler {
 			event.setCanceled(true);
 			event.setSwingHand(false);
 		}
+	}
+
+	@SubscribeEvent
+	public static void onChunkUnload(ChunkEvent.Unload event) {
+		if (event.getLevel().isClientSide()) {
+			ChunkPos pos = event.getChunk().getPos();
+
+			CameraViewAreaExtension.onChunkUnload(pos.x, pos.z);
+		}
+	}
+
+	@SubscribeEvent
+	public static void onRenderFramePost(RenderFrameEvent.Post event) {
+		Minecraft mc = Minecraft.getInstance();
+		Player player = mc.player;
+
+		if (player == null || CameraController.FRAME_CAMERA_FEEDS.isEmpty() || !ConfigHandler.SERVER.frameFeedViewingEnabled.get())
+			return;
+
+		ProfilerFiller profiler = Profiler.get();
+		Map<GlobalPos, CameraFeed> activeFrameCameraFeeds;
+		//+1 helps to reduce stuttering when many frames are active at once
+		double feedsToRender = CameraController.FRAME_CAMERA_FEEDS.size() + 1;
+		double fpsCap = ConfigHandler.CLIENT.frameFeedFpsLimit.get();
+		double currentTime = GLFW.glfwGetTime();
+		double frameInterval = 1.0D / fpsCap;
+		double activeFramesPerMcFrame = Mth.ceil((fpsCap * feedsToRender) / mc.getFps());
+
+		if (fpsCap < 260.0D) {
+			activeFrameCameraFeeds = new HashMap<>();
+
+			for (Entry<GlobalPos, CameraFeed> cameraView : CameraController.FRAME_CAMERA_FEEDS.entrySet()) {
+				double timeBetweenFrames = frameInterval / feedsToRender;
+				double lastActiveTime = cameraView.getValue().lastActiveTime().get();
+
+				if (currentTime < lastActiveTime + frameInterval || currentTime < CameraController.lastFrameRendered + timeBetweenFrames || activeFramesPerMcFrame-- <= 0)
+					continue;
+
+				cameraView.getValue().lastActiveTime().set(currentTime);
+				activeFrameCameraFeeds.put(cameraView.getKey(), cameraView.getValue());
+			}
+		}
+		else
+			activeFrameCameraFeeds = CameraController.FRAME_CAMERA_FEEDS;
+
+		if (activeFrameCameraFeeds.isEmpty())
+			return;
+
+		CameraController.lastFrameRendered = currentTime;
+		profiler.push("gameRenderer");
+		profiler.push("securitycraft:frame_level");
+
+		Level level = player.level();
+		DeltaTracker partialTick = event.getPartialTick();
+		Camera camera = mc.gameRenderer.getMainCamera();
+		Entity oldCamEntity = mc.cameraEntity;
+		Window window = mc.getWindow();
+		int oldWidth = window.getWidth();
+		int oldHeight = window.getHeight();
+		List<SectionRenderDispatcher.RenderSection> oldVisibleSections = mc.levelRenderer.visibleSections.clone();
+		int oldServerRenderDistance = mc.options.serverRenderDistance;
+		int newFrameFeedViewDistance = CameraController.getFrameFeedViewDistance(null);
+		double oldX = player.getX();
+		double oldXO = player.xOld;
+		double oldY = player.getY();
+		double oldYO = player.yOld;
+		double oldZ = player.getZ();
+		double oldZO = player.zOld;
+		float oldXRot = player.getXRot();
+		float oldXRotO = player.xRotO;
+		float oldYRot = player.getYRot();
+		float oldYRotO = player.yRotO;
+		float oldEyeHeight = camera.eyeHeight;
+		float oldEyeHeightO = camera.eyeHeightOld;
+		CameraType oldCameraType = mc.options.getCameraType();
+		Entity securityCamera = new Marker(EntityType.MARKER, level); //A separate entity is used instead of moving the player to allow the player to see themselves
+		Frustum playerFrustum = mc.levelRenderer.getFrustum(); //Saved once before the loop, because the frustum changes depending on which camera is viewed
+
+		mc.gameRenderer.setRenderBlockOutline(false);
+		mc.gameRenderer.setRenderHand(false);
+		mc.gameRenderer.setPanoramicMode(true);
+		mc.options.setServerRenderDistance(newFrameFeedViewDistance);
+		window.setWidth(100);
+		window.setHeight(100); //Different width/height values seem to have no effect, although the ratio needs to be 1:1
+		mc.options.setCameraType(CameraType.FIRST_PERSON);
+		camera.eyeHeight = camera.eyeHeightOld = player.getDimensions(Pose.STANDING).eyeHeight();
+		mc.renderBuffers().bufferSource().endBatch(); //Makes sure that previous world rendering is done
+
+		for (Entry<GlobalPos, CameraFeed> cameraView : activeFrameCameraFeeds.entrySet()) {
+			GlobalPos cameraPos = cameraView.getKey();
+
+			if (cameraPos.dimension().equals(level.dimension())) {
+				BlockPos pos = cameraPos.pos();
+
+				if (level.getBlockEntity(pos) instanceof SecurityCameraBlockEntity be) {
+					if (!isFrameInFrustum(cameraPos, playerFrustum))
+						continue;
+
+					CameraFeed feed = cameraView.getValue();
+					RenderTarget frameTarget = feed.renderTarget();
+					Vec3 cameraEntityPos = new Vec3(pos.getX() + 0.5D, pos.getY() - player.getDimensions(Pose.STANDING).eyeHeight() + 0.5D, pos.getZ() + 0.5D);
+					float cameraXRot = be.getDefaultXRotation();
+					float cameraYRot = be.getDefaultYRotation(be.getBlockState().getValue(SecurityCameraBlock.FACING)) + (float) Mth.lerp(partialTick.getGameTimeDeltaPartialTick(false), be.getOriginalCameraRotation(), be.getCameraRotation()) * Mth.RAD_TO_DEG;
+
+					securityCamera.setPos(cameraEntityPos);
+					mc.setCameraEntity(securityCamera);
+					securityCamera.setXRot(cameraXRot);
+					securityCamera.setYRot(cameraYRot);
+					CameraController.currentlyCapturedCamera = cameraPos;
+					mc.levelRenderer.visibleSections.clear();
+					mc.levelRenderer.visibleSections.addAll(feed.visibleSections());
+					SecurityCraftClient.INSTALLED_IUM_MOD.switchToEmptyRenderLists();
+					profiler.push("securitycraft:discover_frame_sections");
+					CameraController.discoverVisibleSections(cameraPos, newFrameFeedViewDistance, feed);
+					profiler.popPush("securitycraft:bind_frame_target");
+					profiler.pop();
+					RenderSystem.getDevice().createCommandEncoder().clearColorAndDepthTextures(frameTarget.getColorTexture(), 0, frameTarget.getDepthTexture(), 1.0);
+					mc.gameRenderer.renderLevel(DeltaTracker.ONE);
+					SecurityCraftClient.INSTALLED_IUM_MOD.switchToPreviousRenderLists();
+					profiler.push("securitycraft:apply_frame_frustum");
+
+					Frustum frustum = LevelRenderer.offsetFrustum(mc.levelRenderer.getFrustum()); //This needs the frame's newly calculated frustum, so it needs to be queried from inside the loop
+
+					if (be.shouldRotate() || feed.visibleSections().isEmpty() || CameraController.FEED_FRUSTUM_UPDATE_REQUIRED.contains(cameraPos)) {
+						CameraController.FEED_FRUSTUM_UPDATE_REQUIRED.remove(cameraPos);
+						feed.visibleSections().clear();
+
+						for (RenderSection section : feed.sectionsInRange()) {
+							if (frustum.isVisible(section.getBoundingBox()))
+								feed.visibleSections().add(section);
+						}
+					}
+
+					profiler.pop();
+				}
+			}
+		}
+
+		securityCamera.discard();
+		mc.setCameraEntity(oldCamEntity);
+		player.setPosRaw(oldX, oldY, oldZ);
+		player.xOld = player.xo = oldXO;
+		player.yOld = player.yo = oldYO;
+		player.zOld = player.zo = oldZO;
+		player.setXRot(oldXRot);
+		player.xRotO = oldXRotO;
+		player.setYRot(oldYRot);
+		player.yRotO = oldYRotO;
+		camera.setup(level, oldCamEntity == null ? player : oldCamEntity, !mc.options.getCameraType().isFirstPerson(), mc.options.getCameraType().isMirrored(), level.tickRateManager().isEntityFrozen(oldCamEntity) ? 1.0F : partialTick.getGameTimeDeltaPartialTick(true));
+		camera.eyeHeight = oldEyeHeight;
+		camera.eyeHeightOld = oldEyeHeightO;
+		mc.options.setCameraType(oldCameraType);
+		mc.gameRenderer.setRenderBlockOutline(true);
+		mc.levelRenderer.visibleSections.clear();
+		mc.levelRenderer.visibleSections.addAll(oldVisibleSections);
+		window.setWidth(oldWidth);
+		window.setHeight(oldHeight);
+		mc.options.setServerRenderDistance(oldServerRenderDistance);
+		mc.gameRenderer.setRenderHand(true);
+		mc.gameRenderer.setPanoramicMode(false);
+		CameraController.currentlyCapturedCamera = null;
+
+		profiler.pop();
+		profiler.pop();
+	}
+
+	private static boolean isFrameInFrustum(GlobalPos cameraPos, Frustum beFrustum) {
+		for (BlockPos framePos : CameraController.FRAME_LINKS.get(cameraPos)) {
+			if (beFrustum.isVisible(new AABB(framePos)))
+				return true;
+		}
+
+		return false;
 	}
 
 	@SubscribeEvent