+++ src/main/java/net/geforcemods/securitycraft/blocks/KeypadBlock.java
@@ -5,6 +5,7 @@ import net.geforcemods.securitycraft.api.IModuleInventory;
 import net.geforcemods.securitycraft.api.IOwnable;
 import net.geforcemods.securitycraft.api.IPasscodeConvertible;
 import net.geforcemods.securitycraft.api.Owner;
+import net.geforcemods.securitycraft.blockentities.FrameBlockEntity;
 import net.geforcemods.securitycraft.blockentities.KeypadBlockEntity;
 import net.geforcemods.securitycraft.util.BlockUtils;
 import net.geforcemods.securitycraft.util.PlayerUtils;
@@ -151,8 +152,16 @@ public class KeypadBlock extends DisguisableBlock {
 
 		@Override
 		public boolean protect(Player player, Level level, BlockPos pos) {
-			Owner owner = ((IOwnable) level.getBlockEntity(pos)).getOwner();
+			FrameBlockEntity be = (FrameBlockEntity) level.getBlockEntity(pos);
 
+			if (be.getCurrentCamera() != null) {
+				PlayerUtils.sendMessageToPlayer(player, Utils.localize(SCContent.FRAME.get().getDescriptionId()), Utils.localize("messages.securitycraft:frame.cannotConvert"), ChatFormatting.RED);
+				return false;
+			}
+
+			Owner owner = be.getOwner();
+
+			be.dropAllModules();
 			level.setBlockAndUpdate(pos, SCContent.KEYPAD.get().defaultBlockState().setValue(KeypadBlock.FACING, level.getBlockState(pos).getValue(FrameBlock.FACING)).setValue(KeypadBlock.POWERED, false));
 			((IOwnable) level.getBlockEntity(pos)).setOwner(owner.getUUID(), owner.getName());
 			return true;