+++ Changelog.md
@@ -1,3 +1,12 @@
+--------------------------Changelog for the upcoming version of SecurityCraft--------------------------
+
+- New: Cameras can now be viewed on Frames within the world (live camera feeds)[1]
+- New: Several server and client config settings to control chunk loading in camera feeds
+- Change: Camera model animations are now synchronized between players
+- Fix: The display of items in the SC Manual can change too fast in certain situations
+
+[1] Note: Frame blocks that already existed in the world prior to this update will lose their owner. These frames can be broken by anyone so they can be placed down with the proper owner again.
+
 --------------------------Changelog for v1.9.12-beta1 of SecurityCraft--------------------------
 
 - Misc.: The minimum required NeoForge version is 21.5.11-beta